+++ NEWS.md
@@ -1,5 +1,6 @@
 # phenoptr 0.2.6.9000
 
+- Add `make_ppp` function to create marked point patterns.
 - `parse_phenotypes` makes nicer names for expression-based phenotypes.
 - Add `phenotype_columns` in support of akoyabio/phenoptrReports#36.
 - Add `read_field_info` and `get_map_path` functions.